temp

+++ src/main/java/com/teamso/flight_reservation_system/service/SeatService.java
@@ -4,10 +4,13 @@ import com.teamso.flight_reservation_system.dto.SeatDto;
 import com.teamso.flight_reservation_system.entity.Flight;
 import com.teamso.flight_reservation_system.entity.Seat;
 import com.teamso.flight_reservation_system.repository.SeatRepository;
+import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.util.Optional;
+import java.util.concurrent.atomic.LongAdder;
 
 @Service
 public class SeatService {
@@ -22,7 +25,7 @@ public class SeatService {
     }
 
     @Transactional
-    public Seat add(Long flightId, SeatDto seatDto) {
+    public Seat create(Long flightId, SeatDto seatDto) throws IllegalArgumentException {
         Flight flight = flightService.getById(flightId)
                 .orElseThrow(() -> new IllegalArgumentException("Flight not found"));
 
@@ -36,7 +39,7 @@ public class SeatService {
     }
 
     @Transactional
-    public Seat update(Long flightId, Long seatId, SeatDto seatDto) {
+    public Seat update(Long flightId, Long seatId, SeatDto seatDto) throws IllegalArgumentException {
         Flight flight = flightService.getById(flightId)
                 .orElseThrow(() -> new IllegalArgumentException("Flight not found"));
 
@@ -53,7 +56,16 @@ public class SeatService {
     }
 
     @Transactional
-    public void delete(Long flightId, Long seatId) {
+    public Seat updateIsPurchased(Long id, boolean isPurchased) throws IllegalArgumentException {
+        Seat seat = seatRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Seat not found"));
+
+        seat.setPurchased(isPurchased);
+        return seatRepository.save(seat);
+    }
+
+    @Transactional
+    public void delete(Long flightId, Long seatId) throws IllegalArgumentException, IllegalStateException {
         Flight flight = flightService.getById(flightId)
                 .orElseThrow(() -> new IllegalArgumentException("Flight not found"));
 
@@ -64,6 +76,14 @@ public class SeatService {
             throw new IllegalArgumentException("Seat does not belong to the specified flight");
         }
 
+        if(seat.isPurchased()) {
+            throw new IllegalStateException("Seat has been purchased, therefore it cannot be deleted");
+        }
+
         seatRepository.delete(seat);
+    }
+
+    public Optional<Seat> getById(Long id) {
+        return seatRepository.findById(id);
     }
 }
